build: Build a tar.gz package along with ordinary installer
The standard PluginInstall has a nasty bug which uses the source
path as part of installation path. Remove, and use a simple
OpenCPN.app prefix

The loader expects the apple tarball to have paths like:

  - OpenCPN.app/Contents/PlugIns: .dylib plugin file, binaries
    and helper libs.
  - OpenCPN.app/Contents/Resources: .lproj gettext translations
    directories.

The prefix is discarded and could be anything.

Let "make package" build both the traditional installer and the
new installation tarball. Remove references to other formats,
notably rpm.