ci: Enhance the cloudsmith deployment (#73).
New logic added to deploy to different directories depending on
if the commit is tagged or not. Tagged commits goes to a stable
repo, untagged to the unstable one.

Here is also new support for defining the repositories in the
environment so that downstreams can deploy to their own repos
without any code changes.